requirements.txt
django>=5.0,<6.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
requests>=2.32.3
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
